requirements.txt
asgiref==3.3.1
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework==3.12.2
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
